+++ packages/assets-controllers/src/token-prices-service/index.test.ts
@@ -12,6 +12,8 @@ describe('token-prices-service', () => {
         "resetSupportedNetworksCache",
         "SPOT_PRICES_SUPPORT_INFO",
         "getAssetId",
+        "setNetworkConfigs",
+        "resetNetworkConfigsCache",
       ]
     `);
   });